pyproject.toml
[project]
name = "resume-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dateparser>=1.2.2",
    "pdfplumber>=0.11.7",
    "pyautogui>=0.9.54",
    "pymupdf>=1.26.4",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "selenium>=4.35.0",
    "sentence-transformers>=5.1.0",
    "spacy>=3.8.7",
    "webdriver-manager>=4.0.2",
]
